requirements.txt
tensorflow>=2.12.0
numpy>=1.21.0
matplotlib>=3.5.0
pillow>=8.0.0
scikit-learn>=1.0.0
